refactor(runner): extract ip registry to separate module
- move ip allocation logic to ip-registry.ts (~280 lines)
- keep tap pool logic in tap-pool.ts (~480 lines)
- re-export diagnostic functions from tap-pool.ts for api compatibility
